enabledProducts refresh bug
