+++ api/app.py
@@ -5,8 +5,9 @@ from azure.common import AzureMissingResourceHttpError
 from flask import abort, jsonify, request, Response
 from flask_cors import CORS
 
-from calculators import Blend, Bridge, Optimizer
+from calculators import Blend, Bridge
 from config import Config
+from controllers.optimizer import optimizerRequestHandler
 from controllers.products import products_get
 from util import DatabaseOperations as db
 from util.authentication import authorize
@@ -275,210 +276,5 @@ def sizeStepsRequestHandler():
     return {"size_fractions": [round(num, ROUNDING_DECIMALS) for num in db.getSizeSteps()]}
 
 
-# Optimizer handler. This function takes the HttpRequest
-# object as input and returns the best combination of products
-# aswell as the execution time, the fitness score and the
-# amount of iterations.
-
-# name, products, mass, weights, environmental
-def optimizerRequestHandler(
-    value,
-    blend_name,
-    products,
-    mass_goal,
-    weight_dict,
-    environmental_list,
-    option,
-    max_iterations,
-    size_steps_filter,
-):
-    print("Started optimization request...")
-    error_list = []
-
-    if not products:
-        products = []
-
-        product_names = db.listProducts()
-
-        for name in product_names:
-            products.append(product_names[name])
-
-    if not weight_dict:
-        weight_dict = {"best_fit": 100.0, "cost": 0.0, "co2": 0.0, "mass_fit": 0.0}
-    else:
-        try:
-            sum = 0
-            for key in weight_dict:
-                if (weight_dict[key] > 100) or (weight_dict[key] < 0):
-                    raise ValueError()
-                sum += weight_dict[key]
-            if (sum < 95) or (sum >= 105):
-                raise ValueError()
-        except KeyError:
-            error_list.append("Invalid 'weights' input")
-        except TypeError:
-            error_list.append("Invalid 'weights' input")
-        except ValueError:
-            error_list.append("Invalid 'weights' input")
-
-    if not option:
-        option = AVERAGE_PORESIZE_OPTION
-
-    if not max_iterations:
-        max_iterations = DEFAULT_MAX_ITERATIONS
-    if max_iterations < 0:
-        max_iterations = 0
-
-    if (not size_steps_filter) or (size_steps_filter < 0):
-        size_steps_filter = 0
-    elif size_steps_filter > 1:
-        size_steps_filter = 1
-
-    mode = None
-
-    if option == MAXIMUM_PORESIZE_OPTION:
-        mode = Mode.Maximum_Poresize
-    elif option == PERMEABILITY_OPTION:
-        mode = Mode.Permeability
-    elif option == AVERAGE_PORESIZE_OPTION:
-        mode = Mode.Average_Poresize
-    else:
-        error_list.append("Invalid 'option' input")
-
-    size_steps = db.getSizeSteps()
-
-    if value:
-        try:
-            bridge = Bridge.calculateBridgeDistribution(mode, value, size_steps)
-        except ValueError:
-            error_list.append("Invalid 'value' input")
-        except TypeError:
-            error_list.append("Invalid 'value' input")
-    else:
-        error_list.append("Missing 'value' input")
-
-    metadata = db.getMetadata()
-
-    products = []
-
-    for id, product_dict in metadata.items():
-        product_dict["id"] = id
-        try:
-
-            if environmental_list:
-                if product_dict["enviromental"] in environmental_list:
-                    product_dict["cumulative"] = db.getCumulative(id)
-                    product_dict["distribution"] = db.getDistribution(id)
-                    products.append(product_dict)
-            else:
-                product_dict["cumulative"] = db.getCumulative(id)
-                product_dict["distribution"] = db.getDistribution(id)
-                products.append(product_dict)
-
-        except [TypeError, ValueError, KeyError]:
-            error_list.append("Invalid 'id' input")
-
-    if "Invalid 'weights' input" not in error_list:
-        try:
-            weights = [
-                float(weight_dict["best_fit"]) / 100,
-                float(weight_dict["cost"]) / 100,
-                float(weight_dict["co2"]) / 100,
-                float(weight_dict["mass_fit"]) / 100,
-            ]
-        except KeyError:
-            error_list.append("Invalid 'weights' input")
-        except TypeError:
-            error_list.append("Invalid 'weights' input")
-
-    if error_list:
-        return jsonify({"Error": error_list})
-
-    try:
-        (
-            best,
-            best_fit_score,
-            mass_score,
-            cost_score,
-            co2_score,
-            enviromental_score,
-            exec_time,
-            iterations,
-            fitness,
-        ) = Optimizer.optimize(
-            products,
-            weights,
-            bridge,
-            mass_goal,
-            getMaxAndMinValues(metadata),
-            max_iterations,
-            size_steps_filter,
-        )
-    except Exception as e:
-        import sys, traceback
-
-        traceback.print_exc(file=sys.stdout)
-        return f"Probably invalid inputs! {e}", 400
-
-    mass_sum = 0
-    products = []
-
-    for product in products:
-        product["mass"] = int(best[product["id"]]) * product["sack_size"]
-        mass_sum += product["mass"]
-
-    for product in products:
-        products.append(
-            Product(
-                product["id"],
-                "",
-                float(product["mass"]) / mass_sum,
-                product["cumulative"],
-                product["distribution"],
-            )
-        )
-
-    optimal_cumulative, optimal_distribution = Blend.calculateBlendDistribution(products, size_steps)
-
-    response_dict = {}
-    blend_list = []
-
-    for id in best:
-        if best[id] > 0:
-            blend_list.append({"id": id, "sacks": best[id]})
-
-    response_dict["name"] = blend_name
-    response_dict["products"] = blend_list
-    response_dict["performance"] = {
-        "best_fit": round(best_fit_score, ROUNDING_DECIMALS),
-        "mass_fit": round(mass_score, ROUNDING_DECIMALS),
-        "cost": round(cost_score, ROUNDING_DECIMALS),
-        "co2": round(co2_score, ROUNDING_DECIMALS),
-        "enviromental": round(enviromental_score, ROUNDING_DECIMALS),
-    }
-    response_dict["cumulative"] = [round(num, ROUNDING_DECIMALS) for num in optimal_cumulative]
-    response_dict["distribution"] = [round(num, ROUNDING_DECIMALS) for num in optimal_distribution]
-    response_dict["execution_time"] = round(exec_time, ROUNDING_DECIMALS)
-    response_dict["iterations"] = iterations
-    response_dict["fitness"] = round(fitness, ROUNDING_DECIMALS)
-
-    return response_dict
-
-
 # This function gets all the costs and CO2 values from the
 # available poroducts, and returns the max and min value for both.
-def getMaxAndMinValues(metadata):
-    CO2_list = []
-    Cost_list = []
-    MaxAndMinList = []
-
-    for key in metadata:
-        Cost_list.append(int(metadata[key]["cost"]))
-        CO2_list.append(int(metadata[key]["co2"]))
-
-    MaxAndMinList.append(max(Cost_list))
-    MaxAndMinList.append(min(Cost_list))
-    MaxAndMinList.append(max(CO2_list))
-    MaxAndMinList.append(min(CO2_list))
-
-    return MaxAndMinList
